backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
boto3==1.34.25
python-dotenv==1.0.0
pydantic-settings==2.1.0
supabase==2.11.0
httpx==0.27.2
openai==1.12.0
python-multipart==0.0.6
celery==5.3.6
redis==5.0.1
icalendar==5.0.11
pytz==2023.3.post1
pypdf==4.0.1
python-docx==1.1.0
numpy==1.26.3
websockets==12.0
